requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
docopt==0.6.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-dialogflow==2.41.2
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.31.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gTTS==2.5.4
idna==3.10
Jinja2==3.1.6
Js2Py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pipwin==0.5.2
playsound==1.2.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SpeechRecognition==3.14.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
